requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.3.4
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.30.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.2
django-cloudinary-storage==0.3.0
django-configurations==2.2
django-cors-headers==3.13.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-extra-fields==3.0.2
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.2
packaging==22.0
phonenumbers==8.13.7
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.1
tqdm==4.65.0
twilio==8.0.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
vonage==3.3.0
wcwidth==0.2.6
whitenoise==5.2.0
wrapt==1.15.0
yarl==1.8.2
